Revert unnecessary fixes from PR #10
All tests are now passing!

Reverted commits:

- f56b2e9d8e272c7a9f586a88e7ace557e84868cd
  "Added failing test to show issue with second level cache when IUser
  implementation does not overrides Equals/GetHashCode"

- c47bebabd99098a2fb5a10c63a5067c8262b1e06
  "Fix for second level cache miss when comparing users using whole
  entities, not only the Id when IUser implementation have no
  Equals/GetHashCode overrides"

Rationale:

Equals() and GetHashCode() should always be overridden on entities that
will will be used with NHibernate [1]. So it's reasonable for
Rhino.Security to require the IUser implementation to do this.

The default implementation of Equals() (in System.Object) just compares
the hashes of two different objects. Two instances of the same class,
regardless of their values, will evaluate as unequal. This is not the
desired behavior when working with entities in a database

[1] http://nhibernate.info/blog/2008/09/06/identity-field-equality-and-hash-code.html